Move SendAsEMailActions to frame package
