+++ src/main/java/org/jabref/gui/frame/MainMenu.java
@@ -13,8 +13,6 @@ import javafx.scene.control.SeparatorMenuItem;
 import org.jabref.gui.ClipBoardManager;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.Globals;
-import org.jabref.gui.SendAsKindleEmailAction;
-import org.jabref.gui.SendAsStandardEmailAction;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionFactory;
 import org.jabref.gui.actions.StandardActions;
