+++ src/main/java/org/jabref/gui/frame/SendAsKindleEmailAction.java
@@ -1,5 +1,7 @@
-package org.jabref.gui;
+package org.jabref.gui.frame;
 
+import org.jabref.gui.DialogService;
+import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.l10n.Localization;
