+++ src/main/java/org/jabref/gui/frame/SendAsStandardEmailAction.java
@@ -1,9 +1,11 @@
-package org.jabref.gui;
+package org.jabref.gui.frame;
 
 import java.io.IOException;
 import java.io.StringWriter;
 import java.util.List;
 
+import org.jabref.gui.DialogService;
+import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.bibtex.BibEntryWriter;
