+++ src/main/java/org/jabref/gui/maintable/RightClickMenu.java
@@ -10,14 +10,14 @@ import javafx.scene.control.SeparatorMenuItem;
 import org.jabref.gui.ClipBoardManager;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.LibraryTab;
-import org.jabref.gui.SendAsKindleEmailAction;
-import org.jabref.gui.SendAsStandardEmailAction;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionFactory;
 import org.jabref.gui.actions.StandardActions;
 import org.jabref.gui.edit.CopyMoreAction;
 import org.jabref.gui.edit.EditAction;
 import org.jabref.gui.exporter.ExportToClipboardAction;
+import org.jabref.gui.frame.SendAsKindleEmailAction;
+import org.jabref.gui.frame.SendAsStandardEmailAction;
 import org.jabref.gui.keyboard.KeyBindingRepository;
 import org.jabref.gui.linkedfile.AttachFileAction;
 import org.jabref.gui.linkedfile.AttachFileFromURLAction;
